test: 데이터 파싱 테스트
모임 회원 조회 데이터 파싱 테스트
