+++ data/src/main/resources/event_user_list_success.json
@@ -1,0 +1,10 @@
+{
+  "data": [
+    {
+      "name": "이길동",
+      "gender": "M",
+      "introduction": "성남시 분당구에 사는 직장인입니다."
+    }
+  ],
+  "message": "수정, 조회, 삭제, 등록 성공!"
+}
